scraper/requirements.txt
scrapy
extruct
python-arango
